// apps/ai-dial-admin/src/types/deployments/export.ts
export enum DeploymentExportEntityType {
  MODEL_SERVING = 'MODEL_SERVING',
  MCP_CONTAINER = 'MCP_CONTAINER',
  INTERCEPTOR_CONTAINER = 'INTERCEPTOR_CONTAINER',
  ADAPTER_CONTAINER = 'ADAPTER_CONTAINER',
  IMAGE = 'IMAGE',
}

export enum DeploymentExportComponentType {
  MCP_IMAGE_DEFINITION = 'MCP_IMAGE_DEFINITION',
  ADAPTER_IMAGE_DEFINITION = 'ADAPTER_IMAGE_DEFINITION',
  INTERCEPTOR_IMAGE_DEFINITION = 'INTERCEPTOR_IMAGE_DEFINITION',
  MCP_DEPLOYMENT = 'MCP_DEPLOYMENT',
  ADAPTER_DEPLOYMENT = 'ADAPTER_DEPLOYMENT',
  INTERCEPTOR_DEPLOYMENT = 'INTERCEPTOR_DEPLOYMENT',
  NIM_DEPLOYMENT = 'NIM_DEPLOYMENT',
  INFERENCE_DEPLOYMENT = 'INFERENCE_DEPLOYMENT',
}